swapbackend2/requirements.txt
asgiref==3.2.10
attrs==20.3.0
awsebcli==3.19.3
bcrypt==3.2.0
blessed==1.17.12
botocore==1.19.63
CacheControl==0.12.6
cached-property==1.5.2
cachetools==4.2.1
cement==2.8.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
colorama==0.4.3
cryptography==3.2
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.43.0
django-chartjs==2.2.1
django-rest-auth==0.9.5
djangorestframework==3.12.1
djongo==1.3.4
dnspython==2.1.0
docker==4.4.3
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
fcm-django==0.3.7
firebase-admin==4.5.2
future==0.16.0
google-api-core==1.26.0
google-api-python-client==1.12.8
google-auth==1.27.0
google-auth-httplib2==0.0.4
google-cloud-core==1.6.0
google-cloud-firestore==2.0.2
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
gunicorn==20.0.4
httplib2==0.19.0
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.3.3
markdown2==2.3.10
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.0
packaging==20.9
paramiko==2.7.2
pathspec==0.5.9
proto-plus==1.13.0
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyJWT==1.7.1
pymongo==3.11.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-social-auth==0.2.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==0.3.1
rsa==4.7.1
semantic-version==2.5.0
six==1.15.0
sqlparse==0.2.4
termcolor==1.1.0
texttable==1.6.3
uritemplate==3.0.1
urllib3==1.25.11
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.2.0
